requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
librosa
numpy
pillow
python-multipart
soundfile
aiofiles
python-multipart
pydub